+++ src/lib.rs
@@ -1,14 +1,24 @@
 
+#[cfg(any(feature = "line-parser", feature = "line-reader"))]
 pub mod line;
-pub mod parser;
-pub mod ical;
 
+#[cfg(any(feature = "ical-parser", feature = "vcard-parser"))]
+pub mod parser;
 
 pub const PARAM_VALUE_DELIMITER: char = ',';
 pub const VALUE_DELIMITER: char = ':';
 pub const PARAM_DELIMITER: char = ';';
 pub const PARAM_NAME_DELIMITER: char = '=';
 
-pub use ical::IcalParser;
-pub use line::LineReader;
-pub use parser::LineParser;
+
+#[cfg(feature = "ical-parser")]
+pub use parser::ical::IcalParser;
+
+#[cfg(feature = "vcard-parser")]
+pub use parser::vcard::VcardParser;
+
+#[cfg(feature = "line-parser")]
+pub use line::parser::LineParser;
+
+#[cfg(feature = "line-reader")]
+pub use line::reader::LineReader;